ATR-855: regrouped DTOs related to graph events into a separate folder and namespace
