+++ src/Acuminator/Acuminator.Utilities/Roslyn/Semantic/GraphEvents/EventHandlerSignatureType.cs
@@ -1,0 +1,20 @@
+﻿namespace Acuminator.Utilities.Roslyn.Semantic.GraphEvents;
+
+public enum EventHandlerSignatureType : byte
+{
+	/// <summary>
+	/// Method is not an event handler
+	/// </summary>
+	None,
+
+	/// <summary>
+	/// Default signature based on the naming convention
+	/// (e.g., <code>void ARInvoice_RowSelected(PXCache sender, PXRowSelectedEventArgs e)</code>)
+	/// </summary>
+	Default,
+
+	/// <summary>
+	/// Generic signature (e.g., <code>void _(Events.RowSelected&lt;ARInvoice e&gt;)</code>)
+	/// </summary>
+	Generic,
+}
