+++ src/Acuminator/Acuminator.Utilities/Roslyn/Semantic/GraphEvents/EventType.cs
@@ -1,0 +1,31 @@
+﻿namespace Acuminator.Utilities.Roslyn.Semantic.GraphEvents;
+
+/// <summary>
+/// Represents an event type in Acumatica Event Model.
+/// </summary>
+public enum EventType : byte
+{
+	None,
+
+	CacheAttached,
+
+	RowSelecting,
+	RowSelected,
+	RowInserting,
+	RowInserted,
+	RowUpdating,
+	RowUpdated,
+	RowDeleting,
+	RowDeleted,
+	RowPersisting,
+	RowPersisted,
+
+	FieldSelecting,
+	FieldDefaulting,
+	FieldVerifying,
+	FieldUpdating,
+	FieldUpdated,
+
+	CommandPreparing,
+	ExceptionHandling,
+}
